tests/wpt/tests/css/css-ui/parsing/resize-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing resize with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#resize">
<meta name="assert" content="resize supports only the grammar 'none | both | horizontal | vertical'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("resize", "auto");
test_invalid_value("resize", "horizontal vertical");
test_invalid_value("resize", "both 0");
test_invalid_value("resize", "-internal-textarea-auto");
</script>
</body>
</html>
